TrabalhoFinal2/requirements.txt
bidict==0.21.4
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
dnspython==2.2.0
Flask==2.0.3
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-Moment==1.0.2
Flask-SocketIO==5.1.1
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
protobuf==3.19.4
pymongo==4.0.1
python-engineio==4.3.1
python-socketio==5.5.2
requests==2.27.1
six==1.16.0
urllib3==1.26.8
websocket-client==1.2.3
Werkzeug==2.0.3
zope.event==4.5.0
zope.interface==5.4.0
